# docs/swagger.yaml
basePath: /
definitions:
...
  internal_team.AddMemberRequest:
    properties:
      role:
        enum:
        - administrator
        - participant
        - user
        example: participant
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - user_id
    type: object
  internal_team.CreateRequest:
    properties:
      description:
        example: Команда разработки бэкенда
        type: string
      name:
        example: Backend Team
        maxLength: 255
        minLength: 1
        type: string
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - name
    - organization_id
    type: object
  internal_team.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  internal_team.TeamResponse:
    properties:
      data:
        properties:
          created_at:
            type: string
          creator_id:
            type: string
          description:
            type: string
          id:
            type: string
          name:
            type: string
          organization_id:
            type: string
        type: object
    type: object
  internal_team.UpdateRequest:
    properties:
      description:
        example: Обновлённое описание
        type: string
      name:
        example: Backend Team Updated
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API для работы с командой над проектом.
  title: timeManager API
  version: "1.0"
paths:
...
  /api/v1/organizations/{id}/teams:
    get:
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить команды организации
      tags:
      - teams
...
  /api/v1/teams:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Создаёт команду в организации. Создатель должен быть членом организации.
        Поддерживает JSON или multipart (name, description, organization_id + опционально
        avatar).
      parameters:
      - description: Данные команды
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_team.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_team.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать команду
      tags:
      - teams
  /api/v1/teams/{id}:
    delete:
      parameters:
      - description: ID команды
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить команду
      tags:
      - teams
    get:
      parameters:
      - description: ID команды
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_team.TeamResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить команду по ID
      tags:
      - teams
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Поддерживает JSON или multipart (name, description + опционально
        avatar).
      parameters:
      - description: ID команды
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_team.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_team.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить команду
      tags:
      - teams
  /api/v1/teams/{id}/avatar:
    put:
      consumes:
      - multipart/form-data
      description: multipart/form-data с полем avatar. Только член организации может
        изменить.
      parameters:
      - description: ID команды
        in: path
        name: id
        required: true
        type: string
      - description: Файл изображения
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_team.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Установить аватарку команды
      tags:
      - teams
  /api/v1/teams/{id}/members:
    get:
      parameters:
      - description: ID команды
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список участников команды
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Добавлять можно только пользователей, состоящих в организации команды.
      parameters:
      - description: ID команды
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_team.AddMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавить участника в команду
      tags:
      - teams
  /api/v1/teams/{id}/members/{user_id}:
    delete:
      parameters:
      - description: ID команды
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить участника из команды
      tags:
      - teams
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"